+++ scoper/src/main/java/com/brandongogetap/scoper/ChildBuilder.java
@@ -1,0 +1,6 @@
+package com.brandongogetap.scoper;
+
+public interface ChildBuilder<T> {
+
+    Object build(T parent);
+}